pyproject.toml
[tool.poetry]
name = "nett-ef.github.io"
version = "0.0.0"
description = "Scripts for GitHub Pages"
authors = ["Fedor Saveliev <nett_ef@icloud.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.donate-crypto.dependencies]
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"
qrcode = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

